refactor(core): remove CounterpartyService interface, use concrete class
